+++ renderer/resource_loader.js
@@ -22,7 +22,7 @@ export class Loader{
 }
 
 export function loadTextFile(path){
-  // todo use XMLHttpRequest for progress event
+  // todo use XMLHttpRequest for progress event (only needed when bigger files)
   return fetch(path).then(response => {
     if(!response.ok){
       throw new Error("cant load resource")
