+++ utils.js
@@ -1,93 +1,5 @@
 // type checking
-export function classOf(v) {
-  return v.constructor;
-}
-
-export function isString(v){
-  return typeof v === 'string' || v instanceof String;
-}
-
-export function isObject(x){
-  return x.constructor.name === 'Object'
-}
-
-// general stuff
-export function exportAs(obj, name=null){
-  name = name??obj.name;
-  if(name==null){
-    console.error("No name");
-  }
-  window[name] = obj;
-}
-
-export function expectValue(v, name=null){
-  name = name ?? v?.name;
-  if(name==null){
-    throw new ReferenceError("Name not specified!")
-  }
-  if(v==null){
-    throw new ReferenceError(`${name} must not be null!`);
-  }
-  return v;
-}
-
-// math
-export function sum(array, initval=0){
-  return array.reduce((a,b)=>a+b, initval);
-}
-
-export function clamp(v, min, max) {
-  return (min!=null && v<min) ? min : ((max!=null && v>max) ? max : v);
-}
-
-// list stuff
-
-// may modify list inplace, but doesnt have to
-export function iextend(a, b){
-  if(b.length < 32_000){
-    a.push(...b);
-    return a;
-  }
-  if(a.length < b.length/2){
-    // a much smaller than b -  use concat (copy not too expensive)
-    return a.concat(b);
-  }
-  if(b.length < a.length/2){
-    // b much smaller - just use a loop
-    for(const v of b){
-      a.push(v);
-    }
-    return a;
-  }
-  return a.concat(b);
-}
-
-export function extendNullSafe(a, ...args){
-  for(const other of args){
-    if(other==null){ continue; }
-    for(const v of other){
-      if(v==null){ continue; }
-      a.push(v);
-    }
-  }
-  return a;
-}
-
-export function sortCoords(p0, p1){
-  let len = p0.length;
-  if(p0.length != p1.length){
-    throw new RangeError("Arrays must be same len");
-  }
-  for(let i=0;i<len;i++){
-    if(p0[i] > p1[i]){
-    [p0[i], p1[i]] = [p1[i], p0[i]];
-  }
-  }
-  return [p0, p1];
-}
-
-
-
+export * from "./utils/index.js";
 
 // webgl stuff
 export function getGL(){
