+++ utils/file_load.js
@@ -1,0 +1,13 @@
+export function fetchTextFile(path){
+  // todo use XMLHttpRequest for progress event (only needed when bigger files)
+  return fetch(path).then(response => {
+    if(!response.ok){
+      throw new Error("cant load resource")
+    }
+    return response.text();
+  })
+  .catch(reason => {
+    console.error(reason)
+    throw reason;
+  })
+}
